Split Parser from YAMLPath
