+++ bin/eyaml-rotate-keys
@@ -153,7 +153,7 @@ def main():
                         seen_anchors.append(anchor_name)
 
                 log.verbose(
-                    "Decrypting value(s) at {}.".format(yh.str_path(yaml_path))
+                    "Decrypting value(s) at {}.".format(yh.parser.str_path(yaml_path))
                 )
                 yh.publickey = args.oldpublickey
                 yh.privatekey = args.oldprivatekey
