+++ yamlpath/yamlpath.py
@@ -24,12 +24,10 @@ from yamlpath.enums import (
     PathSegmentTypes,
     PathSearchMethods,
 )
+from yamlpath.parser import Parser
 
 class YAMLPath:
     """Query and update YAML data via robust YAML Paths."""
-
-    # Cache parsed YAML Path results across instances to avoid repeated parsing
-    _static_parsings = {}
 
     def __init__(self, logger):
         """Init this class.
@@ -43,6 +41,7 @@ class YAMLPath:
         Raises:  N/A
         """
         self.log = logger
+        self.parser = Parser(logger)
 
     def get_nodes(self, data, yaml_path, mustexist=False, default_value=None):
         """Retrieves zero or more node at YAML Path in YAML data.
@@ -62,7 +61,7 @@ class YAMLPath:
         Raises:
             YAMLPathException when YAML Path is invalid
         """
-        path = self._parse_path(yaml_path)
+        path = self.parser.parse_path(yaml_path)
         if mustexist:
             matched_nodes = 0
             for node in self._get_nodes(data, path):
@@ -76,8 +75,8 @@ class YAMLPath:
             if 1 > matched_nodes:
                 raise YAMLPathException(
                     "Required YAML Path does not match any nodes",
-                    self.str_path(yaml_path),
-                    self.str_path(path)
+                    self.parser.str_path(yaml_path),
+                    self.parser.str_path(path)
                 )
         else:
             for node in self._ensure_path(data, path, default_value):
@@ -111,11 +110,11 @@ class YAMLPath:
         if data is None or yaml_path is None:
             return
 
-        path = self._parse_path(yaml_path)
+        path = self.parser.parse_path(yaml_path)
         if mustexist:
             self.log.debug(
                 "YAMLHelpers::set_value:  Seeking required node at {}."
-                    .format(self.str_path(path))
+                    .format(self.parser.str_path(path))
             )
             found_nodes = 0
             for node in self._get_nodes(data, path):
@@ -128,327 +127,17 @@ class YAMLPath:
             if 1 > found_nodes:
                 raise YAMLPathException(
                     "No nodes matched required YAML Path",
-                    self.str_path(path)
+                    self.parser.str_path(path)
                 )
         else:
             self.log.debug(
                 "YAMLHelpers::set_value:  Seeking optional node at {}."
-                    .format(self.str_path(path))
+                    .format(self.parser.str_path(path))
             )
             for node in self._ensure_path(data, path, value):
                 if node is None:
                     continue
                 self._update_value(data, node, value, format)
-
-    def str_path(self, yaml_path):
-        """Returns the printable, user-friendly version of a YAML Path.
-
-        Positional Parameters:
-          1. yaml_path (any) The YAML Path to convert
-
-        Returns:  (str) The stringified YAML Path
-
-        Raises:  N/A
-        """
-        parsed_path = self._parse_path(yaml_path)
-        add_dot = False
-        ppath = ""
-
-        for (ptype, element_id) in parsed_path:
-            if ptype == PathSegmentTypes.KEY:
-                if add_dot:
-                    ppath += "."
-                ppath += element_id.replace(".", "\\.")
-            elif ptype == PathSegmentTypes.INDEX:
-                ppath += "[" + str(element_id) + "]"
-            elif ptype == PathSegmentTypes.ANCHOR:
-                ppath += "[&" + element_id + "]"
-            elif ptype == PathSegmentTypes.SEARCH:
-                invert, method, attr, term = element_id
-                ppath += (
-                    "["
-                    + str(attr)
-                    + ("!" if invert else "")
-                    + PathSearchMethods.to_operator(method)
-                    + str(term).replace(" ", "\\ ")
-                    + "]"
-                )
-
-            add_dot = True
-
-        return ppath
-
-    def _parse_path(self, yaml_path):
-        r"""Breaks apart a stringified YAML Path into component elements, each
-        identified by its type.  Some non-exhaustive examples of valid YAML
-        Paths include:
-
-           1. bare_top_level_key
-           2. namespaced::top_level_key
-           3. dictionary.key
-           4. dictionary.sub.key
-           5. top_level_list[element_index]
-           6. &top_level_anchor
-           7. aliases[&anchored_element]
-           8. dictionary.'with.dotted.subkey'
-           9. dictionary."with.dotted.subkey"
-          10. dictionary.with\.dottet\.subkey
-          11. complex.structures.with[&many].nested."elements.in.any"[form]
-          12. sensitive::accounts.application.db.users[name=admin].password.encrypted
-          13. sensitive::accounts.application.db.users[name^adm].password.encrypted
-          14. sensitive::accounts.application.db.users[name$in].password.encrypted
-          15. sensitive::accounts.application.db.users[name%dmi].password.encrypted
-          16. sensitive::accounts.application.db.users[name!=admin].password.encrypted
-          17. sensitive::accounts.application.db.users[access_level>3].password.enabled
-          18. sensitive::accounts.application.db.users[access_level<6].password.enabled
-          19. sensitive::accounts.application.db.users[access_level<=5].password.enabled
-          20. sensitive::accounts.application.db.users[access_level<=5].password.[encrypted!^ENC\[]
-
-        Positional Parameters:
-          1. yaml_path (any) The stringified YAML Path to parse
-
-        Returns:  (deque) an empty queue or a queue of tuples, each identifying
-        (type, element) unless yaml_path is already a list or a deque.  If
-        yaml_path is already a list, it is blindly converted into a deque and
-        returned.  If yaml_path is already a deque, it is blindly returned
-        as-is.
-
-        Raises:
-          YAMLPathException when yaml_path is invalid
-        """
-        self.log.debug(
-            "YAMLHelpers::_parse_path:  Evaluating {}...".format(yaml_path)
-        )
-
-        path_elements = deque()
-
-        if yaml_path is None:
-            return path_elements
-        elif isinstance(yaml_path, deque):
-            return yaml_path
-        elif isinstance(yaml_path, list):
-            return deque(yaml_path)
-        elif isinstance(yaml_path, dict):
-            raise YAMLPathException(
-                "YAML paths must be strings, queues, or lists",
-                yaml_path
-            )
-
-        # Check for empty paths
-        if not str(yaml_path).strip():
-            return path_elements
-
-        # Don't parse a path that has already been seen
-        if yaml_path in YAMLPath._static_parsings:
-            return YAMLPath._static_parsings[yaml_path].copy()
-
-        element_id = ""
-        demarc_stack = []
-        seeking_anchor_mark = "&" == yaml_path[0]
-        escape_next = False
-        element_type = PathSegmentTypes.KEY
-        search_inverted = False
-        search_method = PathSearchMethods.EQUALS
-        search_attr = ""
-
-        for c in yaml_path:
-            demarc_count = len(demarc_stack)
-
-            if not escape_next and "\\" == c:
-                # Escape the next character
-                escape_next = True
-                continue
-
-            elif (
-                not escape_next
-                and " " == c
-                and ((1 > demarc_count) or (not demarc_stack[-1] in ["'", '"']))
-            ):
-                # Ignore unescaped, non-demarcated whitespace
-                continue
-
-            elif not escape_next and seeking_anchor_mark and "&" == c:
-                # Found an expected (permissible) ANCHOR mark
-                seeking_anchor_mark = False
-                element_type = PathSegmentTypes.ANCHOR
-                continue
-
-            elif not escape_next and c in ['"', "'"]:
-                # Found a string demarcation mark
-                if 0 < demarc_count:
-                    # Already appending to an ongoing demarcated value
-                    if c == demarc_stack[-1]:
-                        # Close a matching pair
-                        demarc_stack.pop()
-                        demarc_count -= 1
-
-                        # Record the element_id when all pairs have closed
-                        if 1 > demarc_count:
-                            path_elements.append((element_type, element_id))
-                            element_id = ""
-                            element_type = PathSegmentTypes.KEY
-                            continue
-                    else:
-                        # Embed a nested, demarcated component
-                        demarc_stack.append(c)
-                        demarc_count += 1
-                else:
-                    # Fresh demarcated value
-                    demarc_stack.append(c)
-                    demarc_count += 1
-                    continue
-
-            elif not escape_next and "[" == c:
-                if 0 < len(element_id):
-                    # Named list INDEX; record its predecessor element
-                    path_elements.append((element_type, element_id))
-                    element_id = ""
-
-                demarc_stack.append(c)
-                demarc_count += 1
-                element_type = PathSegmentTypes.INDEX
-                seeking_anchor_mark = True
-                search_inverted = False
-                search_method = None
-                search_attr = ""
-                continue
-
-            elif (
-                not escape_next
-                and 0 < demarc_count
-                and "[" == demarc_stack[-1]
-                and c in ["=", "^", "$", "%", "!", ">", "<"]
-            ):
-                # Hash attribute search
-                if "=" == c:
-                    # Exact value match OR >=|<=
-                    element_type = PathSegmentTypes.SEARCH
-
-                    if search_method is PathSearchMethods.LESS_THAN:
-                        search_method = PathSearchMethods.LESS_THAN_OR_EQUAL
-                    elif search_method is PathSearchMethods.GREATER_THAN:
-                        search_method = PathSearchMethods.GREATER_THAN_OR_EQUAL
-                    else:
-                        search_method = PathSearchMethods.EQUALS
-                        search_attr = element_id
-                        element_id = ""
-
-                    continue
-
-                elif "^" == c:
-                    # Value starts with
-                    element_type = PathSegmentTypes.SEARCH
-                    search_method = PathSearchMethods.STARTS_WITH
-                    search_attr = element_id
-                    element_id = ""
-                    continue
-
-                elif "$" == c:
-                    # Value ends with
-                    element_type = PathSegmentTypes.SEARCH
-                    search_method = PathSearchMethods.ENDS_WITH
-                    search_attr = element_id
-                    element_id = ""
-                    continue
-
-                elif "%" == c:
-                    # Value contains
-                    element_type = PathSegmentTypes.SEARCH
-                    search_method = PathSearchMethods.CONTAINS
-                    search_attr = element_id
-                    element_id = ""
-                    continue
-
-                elif ">" == c:
-                    # Value greater than
-                    element_type = PathSegmentTypes.SEARCH
-                    search_method = PathSearchMethods.GREATER_THAN
-                    search_attr = element_id
-                    element_id = ""
-                    continue
-
-                elif "<" == c:
-                    # Value less than
-                    element_type = PathSegmentTypes.SEARCH
-                    search_method = PathSearchMethods.LESS_THAN
-                    search_attr = element_id
-                    element_id = ""
-                    continue
-
-                elif "!" == c:
-                    # Invert the search
-                    search_inverted = True
-                    continue
-
-            elif (
-                not escape_next
-                and 0 < demarc_count
-                and "[" == demarc_stack[-1]
-                and "]" == c
-            ):
-                # Store the INDEX or SEARCH parameters
-                if element_type is PathSegmentTypes.INDEX:
-                    path_elements.append((element_type, int(element_id)))
-                elif element_type is PathSegmentTypes.SEARCH:
-                    # Undemarcate the search term, if it is so
-                    if 0 < len(element_id) and element_id[0] in ["'", '"']:
-                        leading_mark = element_id[0]
-                        if element_id[-1] == leading_mark:
-                            element_id = element_id[1:-1]
-
-                    path_elements.append((
-                        element_type,
-                        [search_inverted,
-                         search_method,
-                         search_attr,
-                         element_id]
-                    ))
-                else:
-                    path_elements.append((element_type, element_id))
-
-                element_id = ""
-                demarc_stack.pop()
-                demarc_count -= 1
-                continue
-
-            elif not escape_next and 1 > demarc_count and "." == c:
-                # Do not store empty elements
-                if 0 < len(element_id):
-                    path_elements.append((element_type, element_id))
-                    element_id = ""
-
-                element_type = PathSegmentTypes.KEY
-                continue
-
-            element_id += c
-            seeking_anchor_mark = False
-            escape_next = False
-
-        # Check for mismatched demarcations
-        if 0 < demarc_count:
-            raise YAMLPathException(
-                "YAML path contains at least one unmatched demarcation mark",
-                yaml_path
-            )
-
-        # Store the final element_id
-        if 0 < len(element_id):
-            path_elements.append((element_type, element_id))
-
-        self.log.debug(
-            "YAMLHelpers::_parse_path:  Parsed {} into:".format(yaml_path)
-        )
-        self.log.debug(path_elements)
-
-        # Cache the parsed results
-        YAMLPath._static_parsings[yaml_path] = path_elements
-        str_path = self.str_path(path_elements)
-        if not str_path == yaml_path:
-            # The stringified YAML Path differs from the user version but has
-            # exactly the same parsed result, so cache it, too
-            YAMLPath._static_parsings[str_path] = path_elements
-
-        return path_elements.copy()
 
     def _get_nodes(self, data, yaml_path):
         """Generates zero or more matching, pre-existing nodes from YAML data
@@ -994,10 +683,10 @@ class YAMLPath:
                     else:
                         restore_path = path.copy()
                         restore_path.appendleft(curref)
-                        restore_path = self.str_path(restore_path)
+                        restore_path = self.parser.str_path(restore_path)
                         throw_element = deque()
                         throw_element.append(curref)
-                        throw_element = self.str_path(throw_element)
+                        throw_element = self.parser.str_path(throw_element)
                         raise YAMLPathException(
                             "Cannot add {} subreference to lists"
                                 .format(str(curtyp))
@@ -1020,10 +709,10 @@ class YAMLPath:
                     else:
                         restore_path = path.copy()
                         restore_path.appendleft(curref)
-                        restore_path = self.str_path(restore_path)
+                        restore_path = self.parser.str_path(restore_path)
                         throw_element = deque()
                         throw_element.append(curref)
-                        throw_element = self.str_path(throw_element)
+                        throw_element = self.parser.str_path(throw_element)
                         raise YAMLPathException(
                             "Cannot add {} subreference to dictionaries".format(
                                 str(curtyp)
@@ -1034,10 +723,10 @@ class YAMLPath:
                 else:
                     restore_path = path.copy()
                     restore_path.appendleft(curref)
-                    restore_path = self.str_path(restore_path)
+                    restore_path = self.parser.str_path(restore_path)
                     throw_element = deque()
                     throw_element.append(curref)
-                    throw_element = self.str_path(throw_element)
+                    throw_element = self.parser.str_path(throw_element)
                     raise YAMLPathException(
                         "Cannot add {} subreference to scalars".format(
                             str(curtyp)
